+++ banyand/tsdb/seriesdb.go
@@ -251,6 +251,7 @@ type SeriesDatabase interface {
 	GetByID(id common.SeriesID) (Series, error)
 	Get(key []byte, entityValues EntityValues) (Series, error)
 	List(ctx context.Context, path Path) (SeriesList, error)
+	SizeOnDisk() int64
 }
 
 type blockDatabase interface {
@@ -271,6 +272,7 @@ type seriesDB struct {
 	seriesMetadata kv.Store
 	l              *logger.Logger
 	segCtrl        *segmentController
+	position       common.Position
 	sync.Mutex
 	sID common.ShardID
 }
@@ -322,8 +324,11 @@ func (s *seriesDB) Get(key []byte, entityValues EntityValues) (Series, error) {
 		}
 		errDecode = s.seriesMetadata.Put(entityKey, encodedData)
 		if errDecode != nil {
+			receivedNumCounter.Inc(1, append(s.position.ShardLabelValues(), "series", "true")...)
 			return nil, errDecode
 		}
+		receivedBytesCounter.Inc(float64(len(entityKey)+len(encodedData)), append(s.position.ShardLabelValues(), "series")...)
+		receivedNumCounter.Inc(1, append(s.position.ShardLabelValues(), "series", "false")...)
 
 		var series string
 		if e := s.l.Debug(); e.Enabled() {
@@ -347,6 +352,10 @@ func (s *seriesDB) Get(key []byte, entityValues EntityValues) (Series, error) {
 	}
 
 	return newSeries(s.context(), seriesID, entityValues.String(), s), nil
+}
+
+func (s *seriesDB) SizeOnDisk() int64 {
+	return s.seriesMetadata.SizeOnDisk()
 }
 
 func encode(seriesID common.SeriesID, evv EntityValues) ([]byte, []byte, error) {
@@ -493,9 +502,10 @@ func (s *seriesDB) Close() error {
 
 func newSeriesDataBase(ctx context.Context, shardID common.ShardID, path string, segCtrl *segmentController) (SeriesDatabase, error) {
 	sdb := &seriesDB{
-		sID:     shardID,
-		segCtrl: segCtrl,
-		l:       logger.Fetch(ctx, "series_database"),
+		sID:      shardID,
+		segCtrl:  segCtrl,
+		l:        logger.Fetch(ctx, "series_database"),
+		position: common.GetPosition(ctx),
 	}
 	o := ctx.Value(optionsKey)
 	var memSize int64