+++ pkg/meter/prom/instruments.go
@@ -28,6 +28,10 @@ func (c *counter) Inc(delta float64, labelValues ...string) {
 	c.counter.WithLabelValues(labelValues...).Add(delta)
 }
 
+func (c *counter) Delete(labelValues ...string) bool {
+	return c.counter.DeleteLabelValues(labelValues...)
+}
+
 type gauge struct {
 	gauge *prometheus.GaugeVec
 }
@@ -40,10 +44,18 @@ func (g *gauge) Add(delta float64, labelValues ...string) {
 	g.gauge.WithLabelValues(labelValues...).Add(delta)
 }
 
+func (g *gauge) Delete(labelValues ...string) bool {
+	return g.gauge.DeleteLabelValues(labelValues...)
+}
+
 type histogram struct {
 	histogram *prometheus.HistogramVec
 }
 
 func (h *histogram) Observe(value float64, labelValues ...string) {
 	h.histogram.WithLabelValues(labelValues...).Observe(value)
+}
+
+func (h *histogram) Delete(labelValues ...string) bool {
+	return h.histogram.DeleteLabelValues(labelValues...)
 }